requirements.txt
Django==2.1.7
django-cors-headers==2.5.1
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.0.1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
pytz==2018.9
